refactor: remove SignIn component and update routing to use Login component
